+++ lib/models/feed/manga_feed.g.dart
@@ -1,0 +1,89 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'manga_feed.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+MangaFeed _$MangaFeedFromJson(Map<String, dynamic> json) => MangaFeed(
+      json['result'] as String?,
+      json['response'] as String?,
+      (json['data'] as List<dynamic>?)
+          ?.map((e) => Data.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      json['limit'] as int?,
+      json['offset'] as int?,
+      json['total'] as int?,
+    );
+
+Map<String, dynamic> _$MangaFeedToJson(MangaFeed instance) => <String, dynamic>{
+      'result': instance.result,
+      'response': instance.response,
+      'data': instance.data,
+      'limit': instance.limit,
+      'offset': instance.offset,
+      'total': instance.total,
+    };
+
+Data _$DataFromJson(Map<String, dynamic> json) => Data(
+      json['id'] as String?,
+      json['type'] as String?,
+      json['attributes'] == null
+          ? null
+          : Attributes.fromJson(json['attributes'] as Map<String, dynamic>),
+    );
+
+Map<String, dynamic> _$DataToJson(Data instance) => <String, dynamic>{
+      'id': instance.id,
+      'type': instance.type,
+      'attributes': instance.attributes,
+    };
+
+Attributes _$AttributesFromJson(Map<String, dynamic> json) => Attributes(
+      json['title'] as String?,
+      json['volume'] as String?,
+      json['chapter'] as String?,
+      json['pages'] as int?,
+      json['translatedLanguage'] as String?,
+      json['uploader'] as String?,
+      json['externalUrl'] as String?,
+      json['version'] as int?,
+      json['createdAt'] as String?,
+      json['updatedAt'] as String?,
+      json['publishedAt'] as String?,
+      json['readableAt'] as String?,
+      (json['relationships'] as List<dynamic>?)
+          ?.map((e) => Relationship.fromJson(e as Map<String, dynamic>))
+          .toList(),
+    );
+
+Map<String, dynamic> _$AttributesToJson(Attributes instance) =>
+    <String, dynamic>{
+      'title': instance.title,
+      'volume': instance.volume,
+      'chapter': instance.chapter,
+      'pages': instance.pages,
+      'translatedLanguage': instance.translatedLanguage,
+      'uploader': instance.uploader,
+      'externalUrl': instance.externalUrl,
+      'version': instance.version,
+      'createdAt': instance.createdAt,
+      'updatedAt': instance.updatedAt,
+      'publishedAt': instance.publishedAt,
+      'readableAt': instance.readableAt,
+      'relationships': instance.relationships,
+    };
+
+Relationship _$RelationshipFromJson(Map<String, dynamic> json) => Relationship(
+      json['id'] as String?,
+      json['type'] as String?,
+      json['related'] as String?,
+    );
+
+Map<String, dynamic> _$RelationshipToJson(Relationship instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'type': instance.type,
+      'related': instance.related,
+    };